--- features/ui/inputHook/formHook.module.scss
@@ -1,5 +1,0 @@
-.formGroup {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-}
