--- features/ui/inputHook/formHook.stories.tsx
@@ -1,15 +1,0 @@
-import { Meta, StoryObj } from "@storybook/react";
-import { FormHook } from "./formHook";
-
-const meta: Meta<typeof FormHook> = {
-  component: FormHook,
-  title: "UI/FormHook",
-};
-
-export default meta;
-
-type Story = StoryObj<typeof FormHook>;
-
-export const Default: Story = {
-  render: () => <FormHook />,
-};
