--- features/ui/inputHookRef/index.ts
@@ -1,2 +1,0 @@
-export { InputHookRef } from "./inputHookRef";
-export { FormHookRef } from "./formHookRef";
