--- features/ui/inputHookZod/index.ts
@@ -1,1 +1,0 @@
-export { InputHookZod } from "./inputHookZod";
